openapi: Allow multiple codegen versions in the same build
We've got a requirement to be able to run codegen with two different
versions of the cli, and that's currently not possible because it's
only possible to declare one cli dependency at a time.

My approach here is to allow the Label of a declared cli jar to be
customisable, and then to allow the rule to be passed a custom Label.

Technically, only the second change is needed as I could declare the
second jar on my own, but it seems reasonable to allow the Label
customisation.