[openapi] Add support for openapi-generator 5.0.0
The CLI of openapi generator changed a bit with the 5.0.0 release.
As mention
[here](https://github.com/OpenAPITools/openapi-generator/blob/150e24dc553a8ea5230ffb938ed3e6020e972faa/docs/global-properties.md#note-on-global-property-declaration)
the global property must now be provided using the `--global-property`
flag instead of the `-D`.

Documentation was updated as well as some tests with specific
openapi-cli versions.

The parsing of the version from the jar might not be resilient (if the
version number contains a `-` for instance) but regular expression are
not available in Starlark.